Uploaded POC files
database.ini holds the info for postgres DB server.  Config.py uses .ini file to config the connection.  Connection.py holds the functions for connecting/disconnecting from the db.  Lists.xlsx is the sample data to upload to the table.  POC.py code that connects to the localhost DB queries the table and then reads in the excel data, inserts into the db, finally running another query to show data was successfully inserted.